requirements.txt
annotated-types==0.7.0
anyio==3.7.1
arrow==1.3.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dateparser==1.2.1
distro==1.9.0
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==2.3.3
fsspec==2025.5.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
ics==0.7.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
openai==1.79.0
openai-whisper==20231117
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
scipy==1.15.3
setuptools==65.5.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
TatSu==5.13.1
tiktoken==0.9.0
torch==2.7.0
tqdm==4.67.1
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3
